requirements.txt
Django==2.2.2
pytest-django==3.5.1
djangorestframework==3.9.4
flake8==3.7.7
psycopg2==2.8.3
